remove jquery.js
 - move API code to api.js
   - some technical debt here because UI code is still in here
 - move util code to util.js
